// src/models/AttendanceStatisticsModel.ts
export class AttendanceStatisticsModel {
    
    workingDays: number;
    workedDays: number;
    workedWithDelay: number;
    absent: number;
    absentWithDuity: number;
    VacDays: number;
    OffDays: number;
    NoTT: number;


    workingDays_Per: number;
    workedDays_Per: number;
    workedWithDelay_Per: number;
    absent_Per: number;
    absentWithDuity_Per: number;
    VacDays_Per: number;
    OffDays_Per: number;
    NoTT_Per: number;

}